cortex/requirements.txt
fastapi
uvicorn
httpx
pypdf
python-multipart
pydantic
sqlalchemy
psycopg2-binary
reportlab
pandas
google-cloud-aiplatform
pgvector
pdf2image
